added hdfs directory access

+++ src/test/scala/net/akmorrow13/endive/featurizers/KmerSuite.scala
@@ -38,7 +38,7 @@ class KmerSuite extends EndiveFunSuite {
   }
 
   sparkTest("should extract kmers using reference and regions") {
-    val trainRDD: RDD[(String, String, ReferenceRegion, Double)] = Preprocess.loadLabels(sc, labelPath)
+    val trainRDD: RDD[(String, String, ReferenceRegion, Int)] = Preprocess.loadLabels(sc, labelPath)
 
     val reference = Sequence(sc.parallelize(Seq(fragment)), sc)
     val sequences: RDD[(ReferenceRegion, String)] = reference.extractSequences(trainRDD.map(_._3))
